【修复】修复cpp抽象bean不包含字段时，生成的bean代码包含两个空构造函数的bug 【特性】go支持多态普通表 【修复】修复java多态表 getAs 泛型函数的bug 【修复】修复python抽象bean不包含字段时， __init__ 为空，导致编译出错的bug.通过 新增 pass 行解决 【修复】修复java可空变量相关生成的bug 【修复】FileCleaner清理目录时，不再删除meta文件 【优化】优化python的Vector2类缩进过多的问题 【更新】更新 README.md

+++ src/Luban.Client/Source/Utils/FileCleaner.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
+using Luban.Common.Utils;
 
 namespace Luban.Client.Common.Utils
 {
@@ -67,7 +69,7 @@ namespace Luban.Client.Common.Utils
                 {
                     s_logger.Trace("remain file:{file}", file);
                 }
-                else
+                else if(!_ignoreFileExtensions.Contains(FileUtil.GetFileExtension(file)))
                 {
                     s_logger.Info("[remove] file: {file}", file);
                     File.Delete(file);
@@ -76,7 +78,7 @@ namespace Luban.Client.Common.Utils
 
             // 清除空目录
             var subDirs = new List<string>(Directory.GetDirectories(dir, "*", SearchOption.AllDirectories));
-            subDirs.Sort((a, b) => -a.CompareTo(b));
+            subDirs.Sort((a, b) => -String.Compare(a, b, StringComparison.Ordinal));
             foreach (var subDir in subDirs)
             {
                 string fullSubDirPath = Path.GetFullPath(subDir);
